Add files via upload
Wazuh is an open-source security platform that provides unified security monitoring, log analysis, intrusion detection, and vulnerability detection for various environments such as cloud, on-premises, and hybrid infrastructures. It is built on top of the OSSEC project and has evolved to include a wide range of features, making it suitable for comprehensive security information and event management (SIEM).